Fix strut reading on x86_64, use long and not CARD32, closing #51.
The bug is that on x86 sizeof(long)==4 and sizeof(CARD32)==4 while
on x86_64 sizeof(long)==8 and sizeof(CARD32)==4.

So if I have pypanel at the bottom of the screen the strut array
is { 0, 0, 0, 24 } each element being of 8 bytes length. But pekwm
assumes them to be of type CARD32 that's only 4 bytes, and it sees only
the first 16 (zero) bytes of the actual array and therefore it doesn't
properly apply the strut cause windows overlap the panel.

The fix is to replace all the arrays of CARD32 with arrays of long when
working with Xlib. I looked through pekwm sources and found only two places
where CARD32 arrays are used.