This should probably fix #38 (vnc issue).
Longer explanation (as far as I understand it):
A client that uses the "Locally Active" input model (see ICCCM 2.0 §4.1.7)
may delegate its input focus to one of its subwindows. If this happens,
a FocusOut event for the top level window is generated but pekwm does not
get the FocusIn event for the subwindow (because it is not managed by
pekwm ?!). WindowManager::handleFocusOutEvent() then gives the input focus
to the root window and thereby removing the input focus from the client
completely.