Fix a wrong memory access.
splitString() is sometimes (in Config.cc) called with a string str from its
second parameter toks. str is therefore a reference to a sting object in
vector but by changing vec later it invalidates this reference.
The solution is to not pass str by reference or change all calls to
splitString(). We do the former.