Close #84, unable to make mplayer window fullscreen.
I installed pekwm-0.1.8 and when I started watching movie with mplayer
and pressed 'f' (make mplayer fullscreen) the size of the window remains the
same, not fullscreen, although its content changed so that I could see the
top-left region of the movie in the small window. It looked like mplayer
incorrectly "thought" the size of the window changed.

The bug seems to be in the code introduced by the revision
7bd7afc250a9f7578de7e38b5ed5925af8860de1.

When debugging I discovered that sometimes ev->value_mask includes
CWStackMode and doesn't include CWY/X/Height/Width. In such case if the
window had been already maximized earlier then the code sets fullscreen
state to STATE_UNSET which is wrong.

The obvious fix (the attached one) is to set fullscreen state only if
one of CWY/X/Height/Width was found in ev->value_mask.

But I guess the actual error is in another place where somebody passes
ev->value_mask with anything unexpected (CWStackMode) althought it
should not.

What do you think?

P.S. The bug appears only when FullscreenDetect is True or unset (since
it defaults to True)