Make templates available in autoproperties only with require syntax.
In order to support having templates overwriting in autoproperties file
a Require section was added just as in the theme file. To enable
templates overwriting add this to your autoproperties file:

 Require {
   Templates = "True"
 }

To support defining title and role in the matching enabling templates
also enables an extended autoproperties syntax.

Example:

 Property = "name,class,role,title,apply on" {
   Sticky = "True"
 }

Enjoy!