Properly handle empty theme directories in menu script
This patch fixes two bugs: First, any arbitrary file in the theme directory
will no longer be treated as a theme--only directories will be listed. Second,
if the directory is empty and '*' is therefore passed as the theme_name, '*'
will not be added to the theme list. Since we use sh and not bash, nullglob is
not an option so a test must be used.