Add two new auto properties AllowedActions DisallowedActions closing #270
Added auto properties to set allowed actions for clients, AllowedActions
and DisallowedActions that can be used on ordinary rules and type rules.

Values for the properties is a space separated list with the following
supported keywords:

  Move, Resize, Iconify, Shade, Stick, MaximizeHorizontal, MaximizeVertical,
  Fullscreen, SetWorkspace and Close

In addtion, the move flag is now honoured when moving and it is now disallowed
for Desktop and Dock type windows.