Apply (modified) patch from wallex, hinting for opacity. Closing #248
This patch (against git-HEAD, 12.03.2010) implements support for setting the opacity hint to windows under various conditions. This hint enables true transparency of windows when a composite manager (eg: xcompmgr) is running alongside PekWM.

Opacity is set in two places: in the autoproperties file (for general windows) and in the config file (for the harbour, menus and workspace indicator).

In the autoproperties file, the property takes the format of: Opacity = "focus, unfocused" (eg: Opacity = "1.0, 0.5").

In the config file, it is defined like this:

In the Screen section, WorkspaceIndicatorOpacity = "value"

In the Harbour section, Opacity = "value"

In the Menu section, FocusOpacity = "value" and UnfocusOpacity = "value" are used.

I know it seems a bit inconsistent, but my original idea was to use a single entry to specify the opacity (ie: like in the autoproperties), however this didn't prove to be the best approach for the config related settings.

Also, all the code is protected by the define OPACITY. I don't know how to mess with configure files, but there needs be added the config option "--enable-opacity" which should define OPACITY when compiling to get it enabled (currently I just manually add the define in pekwm.hh :/).