Workaround: Small windows might cause artifacts when using xcompmgr.
With a small window, like the little statuswindow that is used when a
window is moved around, the border windows might extend beyond the main
window's rectangle. Union the borders' bounding areas together causes the
area beyond the window to be not refreshed properly under xcompmgr. I
suspect a bug in xcompmgr but for the time being clipping the bounding
shape down to the original size fixes this.