Optimise PImage::getMask(.).
A short profiling of pekwm under Xephyr showed that pekwm spends by far the
most time in PImage::getScaledData(). The reason is that PImage::getMask(.)
scales the data first and later createMask(.) actually checks if this
PImage uses an alpha channel at all. If not, the scaling was wasted.
Therefore we add a simple shortcut to getMask(.) and avoid needless calls
to getScaledData().