Bugfix: When a newly mapped window is fullscreen, Frame::fixGeometry() should not resize it to respect the struts.
Alas, the adobe flash browser plugin is not only affected by the
fixGeometry() bug above but it also sets STATE_ABOVE which then lowers
its layer from LAYER_ABOVE_DOCK (assuming FullscreenAbove was set to true)
to LAYER_ONTOP which is below the dock.. and that's the second cause
for #282.

So, hopefully this patch closes #282 and closes #304.