Add ACTION_DEBUG - currently a nop.
I am playing with some ideas for a better debug infrastructure for PekWM,
but sometimes the enum causes conflicts which I'd like to avoid. So the
easy way is to add it to master and forget about it. :) If it isn't needed
in the long run, I'll revert the commit.