Add Debug.{cc,hh} which contain code for logging and debugging purposes.
If HAVE_DEBUG is not defined, the different LOG* macros do nothing but
WARN(), ERR() and ERR_IF() log to std::cerr.

If HAVE_DEBUG is defined, all macros log to RAM and std::cerr by
default. To enable logging to a file a logfile has to be specified (the
default logfile is /dev/null) and enabled ("debug enable logfile"). The
in-memory logging is always done but only the last thirty-two messages
are kept. You can change the number of in-memory messages with
"debug maxmsgs <nr>".