Fix GotoClientID action (closes #235).
Short explanation:
During the time between gotoClient()'s call to giveInputFocus() and
the processing of the FocusIn event generated by this call other events
might cause another call to giveInputFocus() that will override the
first one. These might stem from another FocusIn event or (quite
reliably :) from one of the unmap events caused by the workspace change.
By moving the client's frame to the beginning of the mru list PekWM
will pick the right frame.
This is the least intrusive patch I could come up with. If it is not
enough, we can flush the queue and then ignore things till the last
FocusIn event is reached.