Remove reserved name array and related code from MenuHandler.
MenuHandler::MENU_NAMES_RESERVED was not correctly sorted, so that
the call to std::binary_search(.) lead to undefined behaviour. But since
MenuHandler::reloadStandaloneMenus(.) checks _menu_map for the menu name,
calling MenuHandler::isReservedName(.) was unnecessary to begin with.
This does not fix the double creation of RootMenu and WindowMenu.