Be less restrictive about geometry changes via ConfigureRequest.
Some programs use ConfigureRequest to allow the window being moved with
the mouse. In this case restricting the window to be completely within the
screen is a bit counter-intuitive. But I still want to avoid a misguided
program to place the window completely outside the screen, so leave a
three pixel border inside the screen. After some testing we can adjust
the amount of visible area or change/remove the function.