Don't keep the pixmaps for decor borders and title.
I was wondering why pekwm uses so many pixmaps (with just moderate
usage it was over 600 on my system) and I think the main reason is
that it keeps the pixmaps for all the borders and titles.
Changing the pixmaps after setting them as the background for a
window is probably not even well defined. But the impact of not
keeping them is probably not as bad as one might expect since
pekwm is caching the last few pixmaps already.