Try to improve focus handling in sloppy focus mode
The WorkspaceIndicator and Window menu displayed when focusing next/prev
frames are now being marked to skip the upcoming Enter event if they
are entered. (Leave events are ignored)

This should improve the focus situation when having the cursor in the
center of the screen switching workspace or togging through clients.