Delete unused _loaded fields in theme code
_loaded field is used in destructor to see if unload() should be called,
but in many classes this field is not even initialized (risk of calling
unload when it's not needed), or not set when load() is called.

One way to clean it up is follow the main Theme class: unload in load()
if already loaded, and set _loaded properly. But I think it's overkill
because that's already handled by Theme::load() and Theme::unload().

So this commit just deletes all these useless _loaded fields, saving a
tiny bit of memory. Destructors no longer calls unload() under the
assumption that Theme::~Theme() will call unload() which unloads
everything else already.