Update global keys to use Ctrl Mod4 chains instead of Ctrl Mod1
This is the second point of issue #56. But instead of selectively
switching a few Ctrl-Mod1 chain to Ctrl-Mod4, I think that would just
cause more confusion. Ctrl-Mod1 chain does cause trouble (at least the
Ctrl-Mod1-S in emacs).

I've been using these new keybindings for a couple days. Not so
bad. Losing the ability to use the right Alt (with right hand) could be
a problem to some people. But I have personaly not seen a problem with
it.