Add new parameter "keepfocus" in SendToWorkspace (#75)
The new parameter is meant to be used in the "SendToWorkspace;
GotoWorkspace" command sequence. It makes more sense [1] that the window
you are moving remains focused and on top in the new workspace, so that
you can continue to do whatever you want with it.

This also makes it possible to move the window across multiple
workspaces fast (otherwise you would need find and focus the window
before moving again).

[1] which is also why I've updated data/keys to always use "keepfocus"
in combination with GotoWorkspace.