Lower fullscreen windows when raising a window (#86)
If we raise a window and there is at least one fullscreen window
active (and FullscreenAbove is set to True), the stacking order dictates
that the window is hidden by the fullscreen one. Which means, we
practically don't really raise anything.

To get around that, if we find any fullscreen window in the same
workspace, temporarily lower it. This gives the window in question a
chance to actually show up. When the former fullscreen is raised, it'll
be back at LAYER_ABOVE_DOCK.

This also reverts and redoes 974254f (Don't let new client steal focus
from fullscreen window, 2021-04-04). Since we can show above fullscreen
now (without iconifying it which was one of the objections), there's no
reason to suppress focus and hide a new window. Let it shine.

PWinObj::_fullscreen is removed because I don't like the fact that
Client::isFullscreen manages its fullscreen status somewhere else,
leaving _fullscreen unused (and probably confusing to people too)