Iconify fullscreen windows when raising a window (#86)
If we raise a window and there is at least one fullscreen window active,
the stacking order dictates that the window is hidden by the fullscreen
one. Which means, we practically don't really raise anything.

To get around that, if we find any mapped fullscreen window [1], iconify
it. This gives the window in question a chance to actually show
up. Maximized windows at LAYER_ONTOP is not handled yet, but then you
could have near maximized window at ontop that practically hides
everything anyway, so that's on you.

The change here means any one calling raise() could potentially iconify
fullscreen windows, including by accident.

In reality the only time this could happen is when a fullscreen window
is already active and you have to use keyboard to switch away from
it. Any other action that leads to raising requires focusing on
that (non-fullscreen) window first, which means no fullscreen to iconify
in the first place, so this should be safe.

There are two unrelated changes in this same commit, one is converting
TABs to whitespaces. Another is a comment typo.

[1] The actual test is a bit more lax, in a good way. Any window that
completely covers the window to be raised will be iconfied. So if you
have a very large always-on-top window that hides another window, it
would be iconified too.