Don't let new client steal focus from fullscreen window
When a fullscreen window is active, you can still launch a new window
with keyboard shortcut. The current code will not let the new client's
window show above fullscreen, but it gives focus to the new window
anyway. Which is confusing because the user will interact with the new
window but see the fullscreen one.

One option to do this is iconify fullscreen window. But for some reason
it doesn't work well for me because the focus is still on some other
window instead of the new client's window.

Another, less drastic, option is simply don't give focus to the new
window if the user can't see it. There's a flash when the window is
mapped (I think) and hidden behind fullscreen window when a new client
starts.