Do let clients steal focus from fullscreen sometimes
The commit 974254f (Don't let new client steal focus from fullscreen
window, 2021-04-04) forbids new client from stealing focus from a
fullscreen window because the visible window in that case is still the
fullscreen window.

But that is only true when FullscreenAbove configuration is set to True,
which puts fullscreen windows on their own layer above the normal
one. If this is set to false, a fullscreen window should still be in
normal layer, and the new client should be able to show up. Which means
no need to prevent them from getting focus (and we must not because it
causes the same confusion again the other way around)

wo->getParent()->isFullscreen() is also changed to wo->isFullscreen().
The problem with "wo->isFullscreen()" before was because isFullscreen is
not a virtual function so we're calling PWinObj::isFullscreen() on a
Client object instead of Client::isFullscreen() and the two are not the
same.