Correct maximized state after SetGeometry command
Toggle to maximize a window, execute SetGeometry to move it in a corner,
toggle maximizing again. One would expect the window to go back
maximized because SetGeometry un-maximizes the window, but it's
not. This is the fix.

There are two versions of moveResize(). The one that takes x, y, width
and height leave maximized state handling to the callers (all in
Frame.cc it seems). The other one being updated here is used by
SetGeomtry and MoveResize commands, which I think should update
maximized state correctly.

The second version is also used by setupAPGeometry() and this is where
I'm not sure if clearing maximized state automatically is appropriate...