Log find_program calls and fix missing substitutions
Make it a bit more clear that the awk, sed, sh programs are being
selected during the configuration of pekwm.

It is possible to override by specifying -DAWK=/path/to/awk.

A few instances of awk and sed was not substituted in the scripts,
fix that for a more consistent behavior.