Resize fullscreen/maximized windows on randr changes
When a randr change is detected, currently we try to make sure a window
is still (partly) visible by moving it to where it can be seen. But it
just makes more sense to do a bit more for windows that we calculate
width or height automatically: maximized or fullscreen windows.

For maximized ones, while it's possible for the user to do it manually,
it's quite cumbersome (you have to unmaximize first, then resize by
keyboard; no mouse because you probably don't see the edge). For
fullscreen that's probably impossible.

So, when randr change happens, we now try to resize these windows as
well.

There are two cases this could happen. The first is resolution change
with just one monitor. I've tested this.

The second is when you have two monitor setup with different sizes and
disable the primary (and large) monitor. Suddenly all the windows become
too large for the second and smaller one. This is where I notice this.
But unfortunately I don't have access to this setup at the moment, so I
can't verify it. But it should work the same in theory.