Reduce flicker when switching workspaces
When you have a few windows in each workspace, you may see flicker when
switching workspaces because all the windows in current workspace are
hidden first, showing the background, then windows from the new
workspaces start showing on top.

We could do better by reverse those two operations: let's show windows
from the new workspace first, then hide ones from the old workspace. It
seems to be better (subjectively of course)

It still depends on stacking order though, so a new window could be
behind the current one, we only show the non-overlapping parts of it,
then show the rest when the current window is unmmaped. Not ideal, but
messing with stacking order is probably trickier to get right.

This assumes that the order of unmap, set NET_CURRENT_DESKTOP and map is
_not_ important, because we now do map, set then umap.