Support specifying harbour's head by (xrandr) name
This is not an option with Xinerama. But with randr we have name on
each monitor, which is visible from xrandr or its GUI equivalents.
Selecting head by name would be easier and friendlier. More
importantly, we can also find out which screen is primary from xrandr.

This patch gets that info from randr. Harbour has a new configuration to
locate head by name. MoveToHead is also improved to accept head name in
addition to number. CurrHeadSelector could be enhanced too, but I'm not
sure if it's really useful