Improve MoveToHead command
The first is related to 624287a (Resize fullscreen/maximized windows
on randr changes - 2022-01-28). The same behavior change is made: if
we move a window to a different head and it's fullscreen (or
maximized) then we make sure it fills the head.

The second is to make MoveToHead accept a direction keyword in
addition to head number. This makes it possible to say "move this
window to the left monitor", which could be bound to a key.