Move grouping drag code into separate event handler
Continue with the event handler conversion, now move the grouping drag
handling into a separate event handler.

Noticable difference to the end user is that the grouping drag will stop
as soon as the client being dragged exits (on the next motion event).

Also, doing a grouping drag from a single client frame releasing the
window on the root window will move the client there, making it behave
as ungrouping clients.