Fix harbour placement on the right edge
Commit b780659 (Fix max line-length and too many space indents,
2022-07-16) factors out the code inside Harbour::placeDockApp() into
placeDockAppX and placeDockAppY but never calls the latter. Later on
placeDockAppY() is deleted in another cleanup commit.

I do not understand how I only noticed this now [1] but with default
harbor setting (on the right edge), all the dock apps are now moved to
the left edge. And this seems to fix this, but to be honest I don't
really know what I'm doing.

[1] well not now, a few months but I was too lazy to go check