pekwm_sys: system information services + XSETTINGS manager
Process to get location and act as a XSETTINGS manager. Updates XSETTINGS and
Xrm resources on daytime change in order to update themes and application
apperances.

Commands are forwarded from pekwm to pekwm_sys process using a pipe.

This change is a lot larger than one would expect given it introduces and
re-uses quite a bit of new abstractions.

In addition, the use of SIGALRM is removed sharing the timeout code between
pekwm_wm and pekwm_sys.