UI/Theme scaling support
Support scaling the UI by a given factor to facilitate using pekwm under
HiDPI configurations.

If the factor is 1x, 2x etc the scaling will be done without any
interpolation and pixels remain sharp.

Xft fonts are dependent on the DPI setting of the screen, Pango Cairo
fonts are scaled using the same factor as the rest of the UI. Bitmap fonts
are scaled.

Enable by setting the Scale configuration in the main configuration.

```
Screen {
	Scale = "2.0"
}
```