pekwm_panel: If for conditionally adding widgets to the panel.
It is now possible to conditionally include a widget in the panel by
using simple If expressions in the configuration.

Example that will render the battery bar only if the battery_count
variable has the value 1 or more.

```
Bar = "battery" {
	If = "%battery_count > 0"
	Size = "Pixels 24"
	Text = "B"
}
```

add condition, preparing for conditional widgets

Add TextFormatObserver, to be used to observe If conditions

Get If working for pekwm_panel widgets

wip