autotools: add dist of pekwm_audio_ctrl.sh.in, fix pekwm_screenshot link
pekwm_audio_ctrl.sh.in was missing in the list of files to include in the
distribution tarball.

pekwm_screenshot was not included in the previous update for LDADD, png
development was missing on the host where it was tested.

Build system test tools when using autotools, first step in supporting
running plux tests with an autotools built development tree.