Add keyboard shortcut to menu items.
For menu items, the first _ in the menu will use the letter after it as a
shortcut for selecting and executing the given menu item.

If more than one entry in the menu share letter (keycode), only the first
item with that letter will have a shortcut active.

Shortcut letters are highlighted with a _ under the letter.